Added commit

+++ src/main/java/ru/netology/javaqa/domjavanetz6/service/StatsService.java
@@ -3,25 +3,25 @@ package ru.netology.javaqa.domjavanetz6.service;
 public class StatsService {
 
     // Метод для расчета суммы всех продаж
-    public int calculateTotalSales(int[] sales) {
+    public int calculateTotalSales(long[] sales) {
         int total = 0;
-        for (int sale : sales) {
+        for (long sale : sales) {
             total += sale;
         }
         return total;
     }
 
     // Метод для расчета средней суммы продаж в месяц
-    public int calculateAverageSales(int[] sales) {
+    public int calculateAverageSales(long[] sales) {
         int total = calculateTotalSales(sales);
         return (int) total / sales.length;
     }
 
     // Метод для подсчета количества месяцев, в которых продажи были ниже среднего
-    public int calculateBelowAverageSales(int[] sales) {
+    public int calculateBelowAverageSales(long[] sales) {
         double average = calculateTotalSales(sales);
         int count = 0;
-        for (int sale : sales) {
+        for (long sale : sales) {
             if (sale < average) {
                 count++;
             }
@@ -30,10 +30,10 @@ public class StatsService {
     }
 
     // Метод для подсчета количества месяцев, в которых продажи были выше среднего
-    public int calculateLiveAboveAverage(int[] sales) {
+    public int calculateLiveAboveAverage(long[] sales) {
         double average = calculateTotalSales(sales);
         int count = 0;
-        for (int sale : sales) {
+        for (long sale : sales) {
             if (sale > average) {
                 count++;
             }
@@ -41,7 +41,7 @@ public class StatsService {
         return count;
     }
     // Метод с минимальными продажами среди просмотренных ранее
-    public int calculateMinSales(int[] sales) {
+    public int calculateMinSales(long[] sales) {
         int minMonth = 0;
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] <= sales[minMonth]) { // значит, в рассматриваемом i-м месяце продаж меньше
@@ -53,7 +53,7 @@ public class StatsService {
     }
 
     // Метод с максимальными продажами среди просмотренных ранее
-    public int calculateMaxSales(int[] sales) {
+    public int calculateMaxSales(long[] sales) {
         int maxMonth = 0;
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] >= sales[maxMonth]) { // значит, в рассматриваемом i-м месяце продаж меньше
